### 2025-05-06 #8170 @editorialbot
[PRE REVIEW]: voice: A Comprehensive R Package for Audio Analysis
**Submitting author:** @filipezabala (<a href="http://orcid.org/0000-0002-5501-0877">Filipe J. Zabala</a>)
**Repository:** https://github.com/filipezabala/voice/
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.4.31
