+++ paper/draft-pdf.yml
@@ -22,7 +22,7 @@ jobs:
       - name: Upload
         uses: actions/upload-artifact@v4
         with:
-          name: paper
+          name: master
           # This is the output path where Pandoc will write the compiled
           # PDF. Note, this should be the same directory as the input
           # paper.md
