### 2025-05-06 #8170 @filipezabala
> 👋 [@filipezabala](https://github.com/filipezabala) - please add a valid [open source license](https://opensource.org/licenses) to the root of your repository in a LICENSE file before we can proceed.

Done!
