### 2025-05-06 #8170 @filipezabala
On Tue, May 6, 2025 at 2:43 PM The Open Journals editorial robot <
***@***.***> wrote:

> *editorialbot* left a comment (openjournals/joss-reviews#8170)
> <https://github.com/openjournals/joss-reviews/issues/8170#issuecomment-2855418392>
>
> *Software report:*
>
> github.com/AlDanial/cloc v 1.98  T=0.06 s (2127.1 files/s, 241956.4 lines/s)
> -------------------------------------------------------------------------------
> Language                     files          blank        comment           code
> -------------------------------------------------------------------------------
> HTML                            10            279            175           4179
> R                               81            742           2067           3394
> Text                             8            210              0            626
> Markdown                         5            270             36            488
> TeX                              1             32              1            271
> Python                           9             59            111            235
> Rmd                              3             49            162            116
> YAML                             2              1              4             29
> -------------------------------------------------------------------------------
> SUM:                           119           1642           2556           9338
> -------------------------------------------------------------------------------
>
> *Commit count by author*:
>
>    360	Filipe Zabala
>    101	filipezabala
>     28	filipe
>     21	fz
>      4	Julio Trecenti
>
