### 2025-05-15 #8170 @filipezabala
Dear Team,

Some improvements have been implemented in the `voice` package vignette. The updated version is available at:
🔗 [https://filipezabala.com/voicegnette/](https://filipezabala.com/voicegnette/)

Please review the changes at your convenience and let me know if you have any feedback or suggestions.

Thank you for your time and support.

Best regards,
fz
