### 2025-06-08 #8170 @nkrusch
👋 @sneakers-the-rat  @expectopatronum @jjatria @msultan14 would you be interested in you reviewing this submission for JOSS?
