### 2025-06-11 #8170 @nkrusch
👋  @bagustris @yiitozer would you be interested in you reviewing this submission for JOSS?
