### 2025-06-18 #8170 @filipezabala
Hello @fernandosola,

would you be interested in reviewing this submission for JOSS?
