### 2025-06-18 #8170 @filipezabala
Hello @jbgb,

would you be interested in reviewing this submission for JOSS?
