### 2025-06-18 #8420 @editorialbot
[REVIEW]: voice: A Comprehensive R Package for Audio Analysis
**Submitting author:** @filipezabala (<a href="http://orcid.org/0000-0002-5501-0877">Filipe J. Zabala</a>)
**Repository:** https://github.com/filipezabala/voice/
**Branch with paper.md** (empty if default branch): master
**Version:** v0.5.5
**Archive:** 10.17605/OSF.IO/K2B73
