### 2025-06-18 #8420 @nkrusch
:wave: @filipezabala, thanks for your support in locating more reviewers. We are now ready to begin the review phase.

At this time, I recommend you check the paper references and update them when appropriate. The [editorial bot comment](https://github.com/openjournals/joss-reviews/issues/8420#issuecomment-2985657531) shows several missing DOIs.

The JOSS [guide for citations](https://joss.readthedocs.io/en/latest/paper.html#citations) explains the basic bibliographic formatting. To add a DOI, add it to the `paper.bib` file.

After edits you can re-check the references through the editorialbot.
