+++ paper/paper.bib
@@ -5,7 +5,8 @@
   author={Van Rossum, Guido and Drake Jr, Fred L},
   year={1995},
   publisher={Centrum voor Wiskunde en Informatica Amsterdam},
-  url={https://ir.cwi.nl/pub/5008}
+  url={https://ir.cwi.nl/pub/5008},
+  doi={https://dl.acm.org/doi/10.5555/869369}
 }
 
 @book{vapnik2000nature,
@@ -14,7 +15,8 @@
   year={2000},
   edition={2nd},
   publisher={{Springer Science \& Business Media}},
-  url={https://link.springer.com/book/10.1007/978-1-4757-3264-1}
+  url={https://link.springer.com/book/10.1007/978-1-4757-3264-1},
+  doi={https://doi.org/10.1007/978-1-4757-3264-1}
 }
 
 @article{boersma2001praat,
@@ -25,7 +27,8 @@
   number={9/10},
   pages={341--347},
   year={2001},
-  url={https://www.fon.hum.uva.nl/paul/papers/speakUnspeakPraat_glot2001.pdf}
+  url={https://www.fon.hum.uva.nl/paul/papers/speakUnspeakPraat_glot2001.pdf},
+  doi={https://doi.org/11245/1.200596}
 }
 
 @article{breiman2001random,
@@ -37,7 +40,8 @@
   pages={5--32},
   year={2001},
   publisher={Springer},
-  url={https://link.springer.com/content/pdf/10.1023/A:1010933404324.pdf}
+  url={https://link.springer.com/content/pdf/10.1023/A:1010933404324.pdf},
+  doi={https://doi.org/10.1023/A:1010933404324}
 }
 
 @article{good2001musicxml,
@@ -48,7 +52,8 @@
   number={113-124},
   pages={160},
   year={2001},
-  url={http://michaelgood.info/publications/music/musicxml-for-notation-and-analysis/}
+  url={http://michaelgood.info/publications/music/musicxml-for-notation-and-analysis/},
+  doi={https://doi.org/10.7551/mitpress/2058.003.0010}
 }
 
 @article{cramer2002origins,
@@ -56,7 +61,8 @@
   author={Cramer, Jan Salomon},
   year={2002},
   publisher={Tinbergen Institute Working Paper},
-  url={http://hdl.handle.net/10419/86100}
+  url={http://hdl.handle.net/10419/86100},
+  doi={https://dx.doi.org/10.2139/ssrn.360300}
 }
 
 @inproceedings{dwork2006differential,
@@ -66,7 +72,8 @@
   pages={1--12},
   year={2006},
   organization={Springer},
-  url={http://www.dbis.informatik.hu-berlin.de/fileadmin/lectures/SS2011/VL_Privacy/Differential_Privacy.pdf}
+  url={http://www.dbis.informatik.hu-berlin.de/fileadmin/lectures/SS2011/VL_Privacy/Differential_Privacy.pdf},
+  doi={https://doi.org/10.1007/11787006_1}
 }
 
 @article{tranter2006overview,
@@ -78,7 +85,8 @@
   pages={1557--1565},
   year={2006},
   publisher={IEEE},
-  url={https://ieeexplore.ieee.org/document/1677976/}
+  url={https://ieeexplore.ieee.org/document/1677976/},
+  doi={https://doi.org/10.1109/TASL.2006.878256}
 }
 
 @article{sueur2008seewave,
@@ -88,7 +96,7 @@
   journal = {Bioacoustics},
   volume = {18},
   pages = {213-226},
-  url = {https://doi.org/10.1080/09524622.2008.9753600}
+  doi = {https://doi.org/10.1080/09524622.2008.9753600}
 }
 
[…]
