### 2025-06-19 2ada32d
Removed 'https://doi.org/' prefix.
