+++ paper/paper.bib
@@ -6,7 +6,7 @@
   year={1995},
   publisher={Centrum voor Wiskunde en Informatica Amsterdam},
   url={https://ir.cwi.nl/pub/5008},
-  doi={https://dl.acm.org/doi/10.5555/869369}
+  doi={10.5555/869369}
 }
 
 @book{vapnik2000nature,
@@ -16,7 +16,7 @@
   edition={2nd},
   publisher={{Springer Science \& Business Media}},
   url={https://link.springer.com/book/10.1007/978-1-4757-3264-1},
-  doi={https://doi.org/10.1007/978-1-4757-3264-1}
+  doi={10.1007/978-1-4757-3264-1}
 }
 
 @article{boersma2001praat,
@@ -28,7 +28,7 @@
   pages={341--347},
   year={2001},
   url={https://www.fon.hum.uva.nl/paul/papers/speakUnspeakPraat_glot2001.pdf},
-  doi={https://doi.org/11245/1.200596}
+  doi={11245/1.200596}
 }
 
 @article{breiman2001random,
@@ -41,7 +41,7 @@
   year={2001},
   publisher={Springer},
   url={https://link.springer.com/content/pdf/10.1023/A:1010933404324.pdf},
-  doi={https://doi.org/10.1023/A:1010933404324}
+  doi={10.1023/A:1010933404324}
 }
 
 @article{good2001musicxml,
@@ -53,7 +53,7 @@
   pages={160},
   year={2001},
   url={http://michaelgood.info/publications/music/musicxml-for-notation-and-analysis/},
-  doi={https://doi.org/10.7551/mitpress/2058.003.0010}
+  doi={10.7551/mitpress/2058.003.0010}
 }
 
 @article{cramer2002origins,
@@ -73,7 +73,7 @@
   year={2006},
   organization={Springer},
   url={http://www.dbis.informatik.hu-berlin.de/fileadmin/lectures/SS2011/VL_Privacy/Differential_Privacy.pdf},
-  doi={https://doi.org/10.1007/11787006_1}
+  doi={10.1007/11787006_1}
 }
 
 @article{tranter2006overview,
@@ -86,7 +86,7 @@
   year={2006},
   publisher={IEEE},
   url={https://ieeexplore.ieee.org/document/1677976/},
-  doi={https://doi.org/10.1109/TASL.2006.878256}
+  doi={10.1109/TASL.2006.878256}
 }
 
 @article{sueur2008seewave,
@@ -96,7 +96,7 @@
   journal = {Bioacoustics},
   volume = {18},
   pages = {213-226},
-  doi = {https://doi.org/10.1080/09524622.2008.9753600}
+  doi = {10.1080/09524622.2008.9753600}
 }
 
 @article{anguera2012speaker,
@@ -109,7 +109,7 @@
   year={2012},
   publisher={IEEE},
   url={https://ieeexplore.ieee.org/document/6135543},
-  doi={https://doi.org/10.1109/TASL.2011.2125954}
+  doi={10.1109/TASL.2011.2125954}
 }
 
 @article{cao2014crema,
@@ -122,7 +122,7 @@
   year={2014},
   publisher={IEEE},
   url={https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4313618/},
-  doi={https://doi.org/10.1109/TAFFC.2014.2336244}
+  doi={10.1109/TAFFC.2014.2336244}
 }
 
 @article{vryzas2018speech,
@@ -155,7 +155,7 @@
   pages={1--15},
   year={2018},
   publisher={Elsevier},
-  doi={https://doi.org/10.1016/j.wocn.2018.07.001}
+  doi={10.1016/j.wocn.2018.07.001}
 }
 
[…]
