### 2025-06-19 #8420 @filipezabala
On Wed, Jun 18, 2025 at 8:00 PM Filipe Zabala ***@***.***>
wrote:

> @editorialbot check references
>
>
> On Wed, Jun 18, 2025 at 6:20 PM Neea ***@***.***> wrote:
>
>> *nkrusch* left a comment (openjournals/joss-reviews#8420)
>> <https://github.com/openjournals/joss-reviews/issues/8420#issuecomment-2985726199>
>>
>> 👋 @filipezabala <https://github.com/filipezabala>, thanks for your
>> support in locating more reviewers. We are now ready to begin the review
>> phase.
>>
>> At this time, I recommend you check the paper references and update them
>> when appropriate. The editorial bot comment
>> <https://github.com/openjournals/joss-reviews/issues/8420#issuecomment-2985657531>
>> shows several missing DOIs.
>>
>> The JOSS guide for citations
>> <https://joss.readthedocs.io/en/latest/paper.html#citations> explains
>> the basic bibliographic formatting. To add a DOI, add it to the paper.bib
>> file.
>>
>> After edits you can re-check the references through the editorialbot.
>>
>> @editorialbot check references
>>
>> —
>> Reply to this email directly, view it on GitHub
>> <https://github.com/openjournals/joss-reviews/issues/8420#issuecomment-2985726199>,
>> or unsubscribe
>> <https://github.com/notifications/unsubscribe-auth/AILM4ZPSY72JE52NITW6A733EHJYTAVCNFSM6AAAAAB7UFXZTOVHI2DSMVQWIX3LMV43OSLTON2WKQ3PNVWWK3TUHMZDSOBVG4ZDMMJZHE>
>> .
>> You are receiving this because you were mentioned.Message ID:
>> ***@***.***>
>>
>
