### 2025-06-19 160eb07
Invalid and missing DOIs resolved.
