+++ paper/paper.bib
@@ -5,8 +5,7 @@
   author={Van Rossum, Guido and Drake Jr, Fred L},
   year={1995},
   publisher={Centrum voor Wiskunde en Informatica Amsterdam},
-  url={https://ir.cwi.nl/pub/5008},
-  doi={10.5555/869369}
+  url={https://dl.acm.org/doi/10.5555/869369}
 }
 
 @book{vapnik2000nature,
@@ -62,7 +61,7 @@
   year={2002},
   publisher={Tinbergen Institute Working Paper},
   url={http://hdl.handle.net/10419/86100},
-  doi={https://dx.doi.org/10.2139/ssrn.360300}
+  doi={10.2139/ssrn.360300}
 }
 
 @inproceedings{dwork2006differential,
@@ -135,7 +134,7 @@
   year={2018},
   publisher={Audio Engineering Society},
   url={https://www.researchgate.net/profile/Nikolaos-Vryzas/publication/326005164_Speech_Emotion_Recognition_for_Performance_Interaction/links/5b97e33b299bf14ad4ce3ee5/Speech-Emotion-Recognition-for-Performance-Interaction.pdf},
-  doi={http://dx.doi.org/10.17743/jaes.2018.0036}
+  doi={10.17743/jaes.2018.0036}
 }
 
 @incollection{vryzas2018subjective,
@@ -144,7 +143,7 @@
   booktitle={Proceedings of the Audio Mostly 2018 on Sound in Immersion and Emotion},
   pages={1--7},
   year={2018},
-  doi={https://dl.acm.org/doi/pdf/10.1145/3243274.3243294}
+  doi={10.1145/3243274.3243294}
 }
 
 @article{jadoul2018introducing,
@@ -315,5 +314,6 @@ Inference}},
   author = {Renfei Mao},
   year = {2025},
   note = {R package version 2.0.0, commit 69d639be86b0cef80d815a73a605ad7f9f5e0ce2},
-  url = {https://github.com/flujoo/gm}
+  url = {https://github.com/flujoo/gm},
+  doi={10.32614/cran.package.gm}
 }
