### 2025-06-22 964efb3
Refactor code structure for improved readability and maintainability
