+++ paper/paper.md
@@ -1,81 +1,0 @@
----
-title: 'voice: A Comprehensive R Package for Audio Analysis'
-tags:
-  - R
-  - Python
-  - voice
-  - mood
-  - emotion
-authors:
-  - name: Filipe Jaeger Zabala
-    orcid: 0000-0002-5501-0877
-    equal-contrib: true
-    affiliation: "1, 2"
-  - name: Giovanni Abrahão Salum
-    orcid: 0000-0002-7537-7289
-    equal-contrib: true
-    affiliation: "1, 3"
-affiliations:
- - name: Graduate Program of Psychiatry and Behavioral Sciences, UFRGS, Brazil
-   index: 1
-   ror: 041yk2d64
- - name: Pontifical Catholic University of Rio Grande do Sul
-   index: 2
-   ror: 025vmq686
- - name: Child Mind Institute, New York, NY 10022, USA
-   index: 3
-   ror: 01bfgxw09  
-date: 05 May 2025
-bibliography: paper.bib
----
-
-## Summary
-
-The `voice` package [@zabala2025voice] for R [@r2024r] is a free, open-source toolkit designed to streamline audio analysis by integrating music theory and advanced computational techniques. It enables researchers to extract, tag, and analyze voice data efficiently, supporting applications such as speech recognition, speaker identification, and mood inference. The package simplifies workflows through three core functions: `extract_features`, `tag`, and `diarize`. By bridging gaps in existing tools like `wrassp` [@wrassp2024winkelmann] and `tuneR` [@tuner2023ligges], `voice` offers a unified solution for audio data analysis.
-
-## Statement of Need
-
-Audio data analysis is complex due to variability in file formats and the lack of integrated tools. While packages like `seewave` [@sueur2008seewave] provide foundational capabilities, they often require specialized knowledge. The `voice` package addresses these challenges by combining existing functionalities with novel features such as *Formant Removals*, which enhance predictive accuracy for tasks like sex classification. 
-
-Its user-friendly design makes it accessible to researchers in linguistics, psychology, and bioacoustics, where audio data remains underutilized. There are currently work fronts in these areas making use of `voice` functionalities. By simplifying the extraction and analysis of audio features, `voice` lowers the barrier to entry for researchers and expands the potential for audio data in scientific studies.
-
-## Features
-
-### Core Functions
-
-1. **`extract_features`**:  
-   Extracts standardized audio features from files (e.g., *F0*, *Formant Dispersion*, *Gain*, *MFCC*), leveraging `wrassp` and `tuneR` while introducing new metrics to capture vocal tract characteristics.
-
-2. **`tag`**:  
-   Attaches summarized audio features to datasets, supporting anonymization and privacy-aware analysis via a *6-number summary* (mean, median, standard deviation, coefficient of variation, interquartile range and median absolute deviation).
-
-3. **`diarize`**:  
-   Identifies speaker segments using Python's `pyannote-audio` [@bredin2019pyannote], generating RTTM files for transcription and analysis.
-
-### Novel Contributions
-
-- **Formant Removals**:  
-  Isolates fundamental frequency (F0) from formants, improving feature interpretability for classification tasks. 
-  
-- **Integration of R and Python**:  
-  Uses `reticulate` [@ushey2023reticulate] to combine R's statistical power with Python's tools.
-
-## Example Applications
-
-### Predicting Sex from Voice
[…]
