+++ man/check_chords.Rd
@@ -27,6 +27,8 @@ M$gain[is.na(M$f0)] <- NA
 # assigning notes
 f0_spn <- assign_notes(M, fmt = 0)
 check_chords(f0_spn, window = 3, try_perm = FALSE)
+\dontrun{
 check_chords(f0_spn, window = 3, try_perm = TRUE)
 check_chords(f0_spn, window = 4, try_perm = TRUE)
+}
 }
