### 2025-07-11 7a0f9d5
Version 0.5.2.
