### 2025-07-11 8466512
Vignette.
