+++ vignettes/old/voicegnette_R.Rmd
@@ -21,26 +21,33 @@ knitr::opts_chunk$set(
 ini0 <- Sys.time()
 ```
 
-## 0. Installation
-https://github.com/filipezabala/voice
+## 0. Basic installation
+```{r, eval=FALSE}
+# Development version from GitHub
+devtools::install_github('filipezabala/voice')
+# Stable from CRAN
+install.packages('voice')
+```
+More details at https://github.com/filipezabala/voice.
 
 
 ## 1. Extract features
-### Load packages and audio files
+### 1.1 Load packages and audio files
 ```{r, message=FALSE, warning=FALSE}
 # packs
 library(voice)
+library(tidyverse)
 
 # get path to audio file
 wavDir <- list.files(system.file('extdata', package = 'wrassp'),
                      pattern = glob2rx('*.wav'), full.names = TRUE)
 ```
 
-### Examples
+### 1.2 Extract features
 ```{r, message=FALSE, warning=FALSE}
 # minimal usage
 M <- voice::extract_features(wavDir)
-M
+glimpse(M)
 ```
 
 
@@ -55,4 +62,401 @@ voice::tag(E)
 
 # canonical data
 voice::tag(E, groupBy = 'subject_id')
+```
+
+
+## 3. Visualization
+### 3.1 Get audio
+```{r, message=FALSE, warning=FALSE}
+url0 <- 'https://github.com/filipezabala/voiceAudios/raw/refs/heads/main/wav/doremi.wav'
+download.file(url0, paste0(tempdir(), '/doremi.wav'), mode = 'wb')
+voice::embed_audio(url0) # See https://github.com/mccarthy-m-g/embedr for more details.
+```
+
+### 3.2 Media data
+```{r, message=FALSE, warning=FALSE}
+M <- voice::extract_features(tempdir())
+summary(M)
+```
+
+### 3.3 Plot
+```{r, message=FALSE, warning=FALSE, fig.width=7.5, fig.height=4}
+voice::piano_plot(M, 0) # f0
+voice::piano_plot(M, 0:1) # f0 + f1
+```
+
+### 3.4 Assign notes
+```{r, message=FALSE, warning=FALSE}
+(f0_spn <- voice::assign_notes(M, fmt = 0, min_points = 22, min_percentile = .85)) # f0
+(f1_spn <- voice::assign_notes(M, fmt = 1, min_points = 22, min_percentile = .85)) # f1
+```
+
+### 3.5 Sheet music
+Must have [MuseScore](https://musescore.org/en) and [gm](https://flujoo.github.io/gm/articles/gm.html).
+
+#### 3.5.1 Notes sequence of f0
+```{r, message=FALSE, warning=FALSE}
+library(gm)
+line_0 <- gm::Line(as.character(f0_spn))
+m0 <- gm::Music() +
+  gm::Meter(4, 4) +
+  line_0
+gm::show(m0, to = c('score', 'audio'))
+```
+
+
+#### 3.5.2 Notes sequences of f0 and f1
+```{r, message=FALSE, warning=FALSE}
+line_0 <- gm::Line(as.character(f0_spn))
+line_1 <- gm::Line(as.character(f1_spn))
+m1 <- gm::Music() +
+  gm::Meter(4, 4) +
+  line_0 + line_1
+gm::show(m1, to = c('score', 'audio'))
+```
+
+<!-- #### 3.6.3 Chords sequence using f0, f1 and f2  -->
+<!-- ```{r, message=FALSE, warning=FALSE} -->
+<!-- line_0 <- gm::Line(as.character(f0_spn)) -->
+<!-- line_1 <- gm::Line(as.character(f1_spn)) -->
[…]
