### 2025-07-12 filipezabala/voice#6 @expectopatronum
Details in the Basic Installation section
Hi,

I am one of the reviewers for [your JOSS submission](https://github.com/openjournals/joss-reviews/issues/8420)! I am happy to review your tool :)

I am currently installing the package and have a few remarks that could make it easier for new R users or people that have not used R in a while (like myself):

First, I think it would be helpful to mention that `devtools` needs to be installed:

```install.packages('devtools')```

Afterwards, when trying to install `voice` via the shown command, I receive `WARNING: Rtools is required to build R packages, but is not currently installed.` (probably because I have a new installation of RStudio and R on my system). I am not sure if it is out of scope for your README, but it could be helpful if someone installs a package using `devtools` for the first time.

Then, there are two links in this section:

* First, to https://filipezabala.com/voicegnette/ which contains many details, but not with respect to the installation (as I would have expected). Also, regarding installation, it links back to the README. Since this does not contain more installation details, I think it deserves its own section, e.g. ``Using the package'' :)
* Second, to https://cran.r-project.org/web/packages/voice/index.html, which also does not contain installation details (at least none that helped me). So this might also fit better into another section. And I noticed that the vignette that is linked on this page seems outdated (but maybe it is not supposed to be the same as the one linked about -- sorry, if I am mistaken).

TL;DR: As a user I would find it helpful to have devtools and Rtools mentioned; and move the two links to different sections.

However, after installing devtools and Rtools, installation succeeded :)

Best regards
Verena
