### 2025-07-12 filipezabala/voice#7 @expectopatronum
Review ignored files (.gitignore)
Hi,

I noticed that some `.DS_Store` files are checked into the repo, which shouldn't be the case. I am not sure about [this log file](https://github.com/filipezabala/voice/blob/master/voice.Rcheck/00check.log) but I assume this should also not be in the repository.

Removing those files and adding them to .gitignore would help keep the repository clean.

Thanks for your work :)

Best regards
Verena (Reviewer for [8420](https://github.com/openjournals/joss-reviews/issues/8420))
