### 2025-07-12 #8420 @expectopatronum
@nkrusch thanks for the reminder!

I have a few general questions:
* "Does the full list of paper authors seem appropriate and complete?" Is it ok that the contributors in the repo differ from the names on the paper? "Julio Trecenti" made 4 commits in the repo, but is not on the paper, "Giovanni Abrahão Salum" is on the paper but I don't see them contributing to the repo.
* If I have questions regarding the content of the paper, should I open an issue in the github repository, or should ask here?
* Regarding the state of the field, is it sufficient to discuss software written in R, or would it be also relevant to discuss Python packages, as (in my experience), audio analysis is done in Python (librosa, ...)?
