### 2025-07-12 filipezabala/voice#6 @filipezabala
Hi, Verena.

Thanks for your support.

1. `install.packages('devtools')` added to [vignette](https://filipezabala.com/voicegnette/).
2. Also added information about Windows and macOS installation.
3. I'm in the process of updating the package on CRAN, updates should be available soon.

Best,
fz
