+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: voice
 Type: Package
-Title: Voice Analysis, Speaker Recognition and Mood Inference via Music Theory
+Title: Speaker Recognition, Voice Analysis and Mood Inference via Music Theory
 Version: 0.5.2
 Date: 2025-07-11
 Authors@R: person("Zabala", "Filipe J.", email = "filipezabala@gmail.com", role = c("cre","aut"))
@@ -13,7 +13,7 @@ Description: Provides tools for audio data analysis, including feature extractio
 Depends: R (>= 4.0.0)
 Imports: arrangements, dplyr, ggplot2, R.utils, reticulate, seewave, tabr, tibble, tidyselect, tuneR, wrassp, zoo, htmltools, httr
 Suggests: gm, knitr, tidyverse
-License: GPL-3 + file LICENSE
+License: GPL-3
 Encoding: UTF-8
 LazyData: true
 RoxygenNote: 7.3.2
