+++ DESCRIPTION
@@ -13,8 +13,8 @@ Description: Provides tools for audio data analysis, including feature extractio
 Depends: R (>= 4.0.0)
 Imports: arrangements, dplyr, ggplot2, R.utils, reticulate, seewave, tabr, tibble, tidyselect, tuneR, wrassp, zoo, htmltools, httr
 Suggests: gm, knitr, tidyverse
-License: GPL-3
 LicenseFile: LICENSE
+License: GPL-3
 Encoding: UTF-8
 LazyData: true
 RoxygenNote: 7.3.2
