+++ R/diarize.R
@@ -11,7 +11,7 @@
 #' pyannote.audio (already listed @ https://raw.githubusercontent.com/filipezabala/voice/master/requirements.txt).
 #' #2. Accept \url{https://hf.co/pyannote/segmentation-3.0} user conditions.
 #' #3 Accept \url{https://huggingface.co/pyannote/speaker-diarization-3.1} user conditions.
-#' #4. Create access token.
+#' #4. Create access token in \url{https://hf.co/settings/tokens}.
 #' @return RTTM files in NIST standard. See 'voice::read_rttm'.
 #' @details When \code{autoDir = TRUE}, the following directories are created: \code{'../mp3'},\code{'../rttm'}, \code{'../split'} and \code{'../musicxml'}. Use \code{getwd()} to find the parent directory \code{'../'}.
 #' @import reticulate
