+++ man/diarize.Rd
@@ -26,7 +26,7 @@ pipeline from pyannote.audio. #1 Install pyannote.audio 3.1 with pip install
 pyannote.audio (already listed @ https://raw.githubusercontent.com/filipezabala/voice/master/requirements.txt).
 #2. Accept \url{https://hf.co/pyannote/segmentation-3.0} user conditions.
 #3 Accept \url{https://huggingface.co/pyannote/speaker-diarization-3.1} user conditions.
-#4. Create access token.}
+#4. Create access token in \url{https://hf.co/settings/tokens}.}
 }
 \value{
 RTTM files in NIST standard. See 'voice::read_rttm'.
