### 2025-07-15 filipezabala/voice#7 @expectopatronum
Hi,

thanks for the fast fix! :) I think this one should also not be in the repo: https://github.com/filipezabala/voice/blob/master/voice.Rproj

Best regards
Verena
