### 2025-07-15 filipezabala/voice#6 @expectopatronum
Thanks! I think the installation instructions are clear / complete now :)

Could some of the vignette content also be added to the README of the repo? I think the package could benefit from a more detailed README as it is often the first contact of new users with a package, and one does not have to search the vignette for basic information about the functionality of a package (e.g. an example with some visualization).
