### 2025-07-15 filipezabala/voice#6 @expectopatronum
Later in the vignette optional installs are mentioned (tidyverse, MuseScore, ...) - it would be nice to have them mentioned as optional installs in the installation section already (might be a personal preference but could make the instructions clearer :)).
