+++ README.md
@@ -4,12 +4,458 @@ General tools for voice analysis.
 
 The `voice` package is being developed to be an easy-to-use set of tools to deal with audio analysis in R. It provides a free and user-friendly toolkit for audio analysis, enabling researchers to extract, tag, and analyze voice data efficiently. It supports the extraction of audio features, enrichment of structured datasets with audio summaries, and automatic identification of spoken segments—while introducing novel features. It also allows audio analysis based on musical theory, associating frequencies with musical notes arranged in a score via [`gm`](https://github.com/flujoo/gm) package. 
 
-## Basic installation
-`devtools::install_github('filipezabala/voice')`
-
-More details may be found at 
-- https://filipezabala.com/voicegnette/
-- https://cran.r-project.org/package=voice 
-
-## Issues
 If you want to contribute, report bugs or request new features, use the 'Issues' tab on Github.
+
+
+## 0. Basic installation
+```{r, eval=FALSE}
+# Development version from GitHub
+install.packages('devtools')
+devtools::install_github('filipezabala/voice')
+
+# Stable version from CRAN
+install.packages('voice')
+```
+
+### 0.1 For Windows Users
+If you're compiling R packages from source, you may need to install [RTools](https://cran.r-project.org/bin/windows/Rtools/), a collection of Windows-specific build tools for R.
+
+### 0.2 For macOS Users
+If you're compiling packages, ensure you have [Xcode Command Line Tools](https://mac.install.guide/commandlinetools/) installed. You also may need [macOS tools](https://cran.r-project.org/bin/macosx/tools/).
+
+```{bash, eval=FALSE}
+# Install Xcode on MacOS
+xcode-select --install
+```
+More details may be found at https://github.com/filipezabala/voice.
+
+## 1. Extract features
+### 1.1 Load packages and audio files
+```{r, message=FALSE, warning=FALSE}
+# packs
+library(voice)
+library(tidyverse)
+
+# get path to audio file
+wavDir <- list.files(system.file('extdata', package = 'wrassp'),
+                     pattern = glob2rx('*.wav'), full.names = TRUE)
+```
+
+### 1.2 Extract features
+```{r, message=FALSE, warning=FALSE}
+# minimal usage
+M <- voice::extract_features(wavDir)
+glimpse(M)
+```
+
+
+## 2. Tag
+```{r, message=FALSE, warning=FALSE}
+# creating Extended synthetic data
+E <- dplyr::tibble(subject_id = c(1,1,1,2,2,2,3,3,3), wav_path = wavDir)
+E
+
+# minimal usage
+voice::tag(E)
+
+# canonical data
+voice::tag(E, groupBy = 'subject_id')
+```
+
+
+## 3. Visualization
+### 3.1 Get audio
+```{r, message=FALSE, warning=FALSE}
+url0 <- 'https://github.com/filipezabala/voiceAudios/raw/refs/heads/main/wav/doremi.wav'
+download.file(url0, paste0(tempdir(), '/doremi.wav'), mode = 'wb')
+voice::embed_audio(url0) # See https://github.com/mccarthy-m-g/embedr for more details.
+```
+
+### 3.2 Media data
+```{r, message=FALSE, warning=FALSE}
+M <- voice::extract_features(tempdir())
+summary(M)
+```
+
+### 3.3 Plot
[…]
