### 2025-07-15 filipezabala/voice#7 @filipezabala
Hi, Verena.

I inserted *.Rproj into .gitgnore, thanks for the suggestion.

Best,
fz
