### 2025-07-15 filipezabala/voice#6 @expectopatronum
Hi,

sorry I forget to mention that, I am on Windows.

Best regards
Verena
