### 2025-07-15 1da7298
Drop voice::embed_audio().
