### 2025-07-15 filipezabala/voice#6 @filipezabala
Hi, Verena.

Testing on Windows:
I encountered this crash when running `voice::embed_audio(url0)` before `M <- voice::extract_features(tempdir())`.

The `voice::embed_audio()` function is primarily useful when working with an R Markdown (.Rmd) file. To clarify this, I've rewritten the README, dropping it. Could you please reboot your RStudio and verify if the new sequence works as expected?

Best,
fz
