### 2025-07-17 filipezabala/voice#10 @expectopatronum
Documentation: Statement of Need & Automatic Tests
Hi,

the last part of my [review](https://github.com/openjournals/joss-reviews/issues/8420) is the documentation.

I am missing two parts:

* The review asks for "A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?" also in the documentation. Since you have already written it in the paper, you could copy it to the Github repo? (I personally do not think it needs to be in the vignette)
* It also asks for "Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?" I don't see either of them. Did you do any testing to see whether the results are correct? (e.g. whether the plots are correct, ...)

Best regards
Verena
