### 2025-07-18 filipezabala/voice#10 @filipezabala
Hi Verena,

Thank you for your feedback on the documentation and testing. Here’s how I’d like to address your points:

#### 1. Statement of Need
The README currently begins with a concise overview of the package’s goals and audience (copied below). Would you like me to:

- Expand it with the more detailed "statement of need" from the paper?
- Rephrase it to align closer with the paper’s language?

Current README intro:

> General tools for voice analysis.
>
> The voice package is being developed to be an easy-to-use set of tools to deal with audio analysis in R. It provides a free and user-friendly toolkit for audio analysis, enabling researchers to extract, tag, and analyze voice data efficiently. It supports the extraction of audio features, enrichment of structured datasets with audio summaries, and automatic identification of spoken segments—while introducing novel features. It also allows audio analysis based on musical theory, associating frequencies with musical notes arranged in a score via [gm](https://github.com/flujoo/gm) package.
>
> If you want to contribute, report bugs or request new features, use the 'Issues' tab on Github.

#### 2. Testing
The package has been tested extensively since 2019, including:

- Real-world applications: Dozens of uses (e.g., sex prediction from voice features, speaker diarization in audiobooks).
- Validation: Results documented in the paper’s "Example Applications" section (e.g., successful tests on open datasets and LibriVox recordings).

To improve transparency, I can add a dedicated "Testing" section to the documentation listing key validation cases.

Would either (or both) of these approaches meet your expectations? I’m happy to adapt based on your suggestions.

Best regards,
fz
