### 2025-07-20 filipezabala/voice#10 @expectopatronum
Hi,

I am sorry for not responding yet - I got really sick last week and it will take a few more days. Sorry for delaying the review, I'll finish it as soon as possible.

Best regards
Verena
