### 2025-07-23 #8420 @nkrusch
@filipezabala, just checking in on status.

Since your latest comment, I see the opened new issue. The review of @expectopatronum is just one check box away from complete. Is it correct to assume, @filipezabala, you are working on that last item?

Regarding,

> please let me know if you’d like me to update the paper directory with the latest changes at this stage, or if further revisions are anticipated.

yes, please merge all requested changes as they relate to the paper and reviews, as we prepare to wrap up this review.
