+++ README.md
@@ -4,6 +4,11 @@ General tools for voice analysis.
 
 The `voice` package is being developed to be an easy-to-use set of tools to deal with audio analysis in R. It provides a free and user-friendly toolkit for audio analysis, enabling researchers to extract, tag, and analyze voice data efficiently. It supports the extraction of audio features, enrichment of structured datasets with audio summaries, and automatic identification of spoken segments—while introducing novel features. It also allows audio analysis based on musical theory, associating frequencies with musical notes arranged in a score via [`gm`](https://github.com/flujoo/gm) package. 
 
+The package has been tested extensively since 2019, including:
+
+- Real-world applications: Dozens of uses, e.g. sex prediction from voice features and speaker diarization in audiobooks.
+- Validation: Successful tests on open datasets and LibriVox recordings.
+
 If you want to contribute, report bugs or request new features, use the 'Issues' tab on Github.
 
 
