### 2025-07-23 filipezabala/voice#10 @filipezabala
Hi Verena,

1. Understood - I’ll keep the statement of need as is in the README.
2. I’ve added some information on how the package was tested.

Best regards,
fz
