+++ paper/paper.bib
@@ -1,0 +1,336 @@
+% BIBLIOGRAPHY -----------------------------------------
+
+@book{van1995python,
+  title={Python reference manual},
+  author={Van Rossum, Guido and Drake Jr, Fred L},
+  year={1995},
+  publisher={Centrum voor Wiskunde en Informatica Amsterdam},
+  doi={10.5555/869369}
+}
+
+@book{vapnik2000nature,
+  title={{The Nature of Statistical Learning Theory}},
+  author={Vapnik, Vladimir},
+  year={2000},
+  edition={2nd},
+  publisher={{Springer Science \& Business Media}},
+  url={https://link.springer.com/book/10.1007/978-1-4757-3264-1},
+  doi={10.1007/978-1-4757-3264-1}
+}
+
+@article{boersma2001praat,
+  title={Praat, a system for doing phonetics by computer},
+  author={Boersma, Paul and van Heuven, Vincent},
+  journal={Glot. Int.},
+  volume={5},
+  number={9/10},
+  pages={341--347},
+  year={2001},
+  url={https://www.fon.hum.uva.nl/paul/papers/speakUnspeakPraat_glot2001.pdf},
+  doi={11245/1.200596}
+}
+
+@article{breiman2001random,
+  title={{Random Forests}},
+  author={Breiman, Leo},
+  journal={{Machine Learning}},
+  volume={45},
+  number={1},
+  pages={5--32},
+  year={2001},
+  publisher={Springer},
+  doi={10.1023/A:1010933404324}
+}
+
+@article{good2001musicxml,
+  title={MusicXML for notation and analysis},
+  author={Good, Michael},
+  journal={The virtual score: representation, retrieval, restoration},
+  volume={12},
+  number={113-124},
+  pages={160},
+  year={2001},
+  url={http://michaelgood.info/publications/music/musicxml-for-notation-and-analysis/},
+  doi={10.7551/mitpress/2058.003.0010}
+}
+
+@article{cramer2002origins,
+  title={{The Origins of Logistic Regression}},
+  author={Cramer, Jan Salomon},
+  year={2002},
+  publisher={Tinbergen Institute Working Paper},
+  url={http://hdl.handle.net/10419/86100},
+  doi={10.2139/ssrn.360300}
+}
+
+@inproceedings{dwork2006differential,
+  title={{Differential Privacy}},
+  author={Dwork, Cynthia},
+  booktitle={International Colloquium on Automata, Languages, and Programming},
+  pages={1--12},
+  year={2006},
+  organization={Springer},
+  url={http://www.dbis.informatik.hu-berlin.de/fileadmin/lectures/SS2011/VL_Privacy/Differential_Privacy.pdf},
+  doi={10.1007/11787006_1}
+}
+
+@article{tranter2006overview,
+  title={An overview of automatic speaker diarization systems},
+  author={Tranter, Sue E and Reynolds, Douglas A},
+  journal={IEEE Transactions on audio, speech, and language processing},
+  volume={14},
+  number={5},
+  pages={1557--1565},
+  year={2006},
+  publisher={IEEE},
+  url={https://ieeexplore.ieee.org/document/1677976/},
+  doi={10.1109/TASL.2006.878256}
+}
+
+@article{sueur2008seewave,
+  title = {Seewave: a free modular tool for sound analysis and synthesis},
+  author = {J. Sueur and T. Aubin and C. Simonis},
+  year = {2008},
+  journal = {Bioacoustics},
+  volume = {18},
+  pages = {213-226},
+  doi = {10.1080/09524622.2008.9753600}
+}
+
+@article{anguera2012speaker,
[…]
