### 2025-07-23 #8420 @filipezabala
@nkrusch, thank you for your message and for following up on this.

I can confirm I've completed the last item from @expectopatronum's review by updating the README with all suggested changes. The paper directory has also been updated with all requested revisions.

Please let me know if you need anything else from my side.

Best regards,
fz
