+++ paper/paper.bib
@@ -5,7 +5,7 @@
   author={Van Rossum, Guido and Drake Jr, Fred L},
   year={1995},
   publisher={Centrum voor Wiskunde en Informatica Amsterdam},
-  doi={10.5555/869369}
+  url={https://dl.acm.org/doi/10.5555/869369}
 }
 
 @book{vapnik2000nature,
