### 2025-07-23 #8420 @filipezabala
Dear @nkrusch,

0. the status after replace with [https://dl.acm.org/doi/{doi}](https://dl.acm.org/doi/%7Bdoi%7D) in the {url} field:

❌ MISSING DOIs

None
❌ INVALID DOIs

None

1. Make a tagged release of your software
```bash
$ git tag -a v0.5.5 -m "Release v0.5.5"
$ git push origin v0.5.5
Enumerating objects: 1, done.
Counting objects: 100% (1/1), done.
Writing objects: 100% (1/1), 165 bytes | 165.00 KiB/s, done.
Total 1 (delta 0), reused 0 (delta 0), pack-reused 0
To github.com:filipezabala/voice.git
 * [new tag]         v0.5.5 -> v0.5.5
$ gh release create v0.5.5 --title "v0.5.5" --notes "JOSS review done"
```
2. Archive the reviewed software in Zenodo (or a similar service)
- https://osf.io/k2b73/files/fd7a247b-0247-48d1-a4c8-e8b63c19facc
3. Check that the archival deposit has the correct metadata. This includes the title (should match the paper title) and author list (should match the paper), and authors' ORCID.
- https://osf.io/k2b73/
4. Please list the DOI of the archived version here.
- 10.17605/OSF.IO/K2B73

Best regards,
fz
