### 2025-07-30 #8420 @filipezabala
Dear @crvernon,

I've updated the references as suggested—capitalizing the 'B' in 'Bayesian' (Line 100) and the 'P' in 'Python' (Line 104).

Thank you for your review and support!
Best regards,
fz
