### 2025-07-30 #8420 @crvernon
🥳 Congratulations on your new publication @filipezabala ! Many thanks to @nkrusch for editing and @expectopatronum, @jbgb, and @fernandosola for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
